requirements.txt
alembic==1.11.1
APScheduler==3.10.1
greenlet==2.0.2
Mako==1.2.4
MarkupSafe==2.1.2
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.12
TgCrypto==1.2.5
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1

python-dotenv~=1.0.0